refactor: isolate consent timestamp clamping
